(Temporarily?) Remove function build against the CRoaring amalgamation
The tellmeall functionality used for print_platform_information
in CRoaring isn't exposed in the header files - this was handled
previously with a custom extra in the amalgamation.py script.

For now, rather than hack in extras in the amalgamation, just
remove the print_information functions that call tellmeall and
focus on getting everything else working.